requirements.txt
asttokens==2.2.1
bandit==1.7.4
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.29
icecream==2.1.3
idna==3.4
mypy-extensions==0.4.3
openssl-python==0.1.1
pathspec==0.10.3
pbr==5.11.0
platformdirs==2.6.0
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
smmap==5.0.0
stevedore==4.1.1
tomli==2.0.1
urllib3==1.26.13
